Add GetOrDefault<T> implementation
GetOrDefault<T> attempts to convert the raw value of the setting into
the requested type using the converters TryConvertFromString method.
If conversion fails, a default value specified by the parameter 'def'
is returned instead. Optionally, this default value is set as the new
value for the setting if 'setDef' is true.